backend/requirements/base.txt
channels==1.1.3
Django==1.11
django-filter==1.0.2
djangorestframework==3.6.2
psycopg2==2.7.1
